// frontend/src/types.ts
export interface Feedback {
    text: string;
    rating: number;
    context: string;
    category: string;
    simplifiedText: string;
}

export enum OpenAILanguage {
    Afrikaans = "af",
    Albanian = "sq",
    Amharic = "am",
    Arabic = "ar",
    Armenian = "hy",
    Assamese = "as",
    Azerbaijani = "az",
    Basque = "eu",
    Belarusian = "be",
    Bengali = "bn",
    Bosnian = "bs",
    Bulgarian = "bg",
    Catalan = "ca",
    Chinese = "zh",
    Croatian = "hr",
    Czech = "cs",
    Danish = "da",
    Dutch = "nl",
    English = "en",
    Estonian = "et",
    Faroese = "fo",
    Finnish = "fi",
    French = "fr",
    Galician = "gl",
    Georgian = "ka",
    German = "de",
    Greek = "el",
    Gujarati = "gu",
    Hebrew = "he",
    Hindi = "hi",
    Hungarian = "hu",
    Icelandic = "is",
    Indonesian = "id",
    Irish = "ga",
    Italian = "it",
    Japanese = "ja",
    Kannada = "kn",
    Kazakh = "kk",
    Khmer = "km",
    Korean = "ko",
    Lao = "lo",
    Latvian = "lv",
    Lithuanian = "lt",
    Macedonian = "mk",
    Malay = "ms",
    Malayalam = "ml",
    Maltese = "mt",
    Marathi = "mr",
    Mongolian = "mn",
    Nepali = "ne",
    Norwegian = "no",
    Oriya = "or",
    Pashto = "ps",
    Persian = "fa",
    Polish = "pl",
    Portuguese = "pt",
    Punjabi = "pa",
    Romanian = "ro",
    Russian = "ru",
    Serbian = "sr",
    Sindhi = "sd",
    Sinhala = "si",
    Slovak = "sk",
    Slovenian = "sl",
    Spanish = "es",
    Swahili = "sw",
    Swedish = "sv",
    Tamil = "ta",
    Telugu = "te",
    Thai = "th",
    Turkish = "tr",
    Ukrainian = "uk",
    Urdu = "ur",
    Uzbek = "uz",
    Vietnamese = "vi",
    Welsh = "cy",
    Xhosa = "xh",
    Zulu = "zu",
  }
  